k-NN function completed

+++ mycode/include/lis2dh.h
@@ -1,1 +1,1 @@
-extern void fetch_and_display(const struct device *sensor, double *xyz);
+extern void fetch_and_display(const struct device *sensor, int *xyz);
